pyproject.toml
[tool.poetry]
name = "aveiro tech city"
version = "0.1.0"
description = ""
authors = ["Mycsina <andremacardoso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
seaborn = "^0.13.0"
ipykernel = "^6.25.2"
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"
questdb = "^1.1.0"
psycopg = "^3.1.12"
influxdb-client = "^1.38.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
